+++ operator/frontend_controller_suite_test.go
@@ -1,4 +1,4 @@
-package controllers
+package operator
 
 import (
 	"context"
