+++ operator/metrics.go
@@ -1,4 +1,4 @@
-package controllers
+package operator
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
