+++ operator/status.go
@@ -1,4 +1,4 @@
-package controllers
+package operator
 
 import (
 	"context"
